requirements.txt
Flask==2.2.2
Flask_Caching==2.0.1
icalendar==4.1.0
python-dotenv==0.21.0
trakt==3.4.0
